+++ internal/service/menu.go
@@ -2,16 +2,16 @@ package service
 
 import (
 	"errors"
-	"hot-coffee/internal/repository"
-	"hot-coffee/models"
+	"frappuccino/internal/repository"
+	"frappuccino/models"
 )
 
 type MenuService interface {
 	CreateMenuItem(item models.MenuItem) (models.MenuItem, error)
 	GetMenuItems() ([]models.MenuItem, error)
-	GetMenuItem(id string) (models.MenuItem, error)
-	UpdateMenuItem(id string, item models.MenuItem) (models.MenuItem, error)
-	DeleteMenuItem(id string) error
+	GetMenuItem(id int64) (models.MenuItem, error)
+	UpdateMenuItem(id int64, item models.MenuItem) (models.MenuItem, error)
+	DeleteMenuItem(id int64) error
 }
 
 type menuService struct {
@@ -29,7 +29,7 @@ func (s *menuService) CreateMenuItem(item models.MenuItem) (models.MenuItem, err
 	if item.Price <= 0 {
 		return models.MenuItem{}, errors.New("price must be positive")
 	}
-	item.ID = generateOrderID()
+	// item.ID = generateOrderID()
 	return s.repo.Create(item)
 }
 
@@ -37,17 +37,29 @@ func (s *menuService) GetMenuItems() ([]models.MenuItem, error) {
 	return s.repo.GetAll()
 }
 
-func (s *menuService) GetMenuItem(id string) (models.MenuItem, error) {
+func (s *menuService) GetMenuItem(id int64) (models.MenuItem, error) {
+	if id == 0 {
+		return models.MenuItem{}, errors.New("id is required")
+	}
 	return s.repo.GetByID(id)
 }
 
-func (s *menuService) UpdateMenuItem(id string, item models.MenuItem) (models.MenuItem, error) {
+func (s *menuService) DeleteMenuItem(id int64) error {
+	if id == 0 {
+		return errors.New("id is required")
+	}
+	return s.repo.Delete(id)
+}
+
+func (s *menuService) UpdateMenuItem(id int64, item models.MenuItem) (models.MenuItem, error) {
 	if id != item.ID {
 		return models.MenuItem{}, errors.New("ID in path doesn't match ID in body")
 	}
+	if item.Name == "" {
+		return models.MenuItem{}, errors.New("name is required")
+	}
+	if item.Price <= 0 {
+		return models.MenuItem{}, errors.New("price must be positive")
+	}
 	return s.repo.Update(id, item)
-}
-
-func (s *menuService) DeleteMenuItem(id string) error {
-	return s.repo.Delete(id)
 }
